requirements.txt
# Optional dependency for generate_image.py
google-genai
